+++ .changeset/shaggy-bulldogs-notice.md
@@ -1,0 +1,5 @@
+---
+'@ultraviolet/ui': patch
+---
+
+Fix right position on popup
