+++ packages/ui/src/components/Popover/__stories__/index.stories.tsx
@@ -8,7 +8,6 @@ export default {
 } as Meta<typeof Popover>
 
 export { Playground } from './Playground.stories'
-export { Controlled } from './Controlled.stories'
 export { Sentiments } from './Sentiments.stories'
 export { Sizes } from './Sizes.stories'
 export { AdvancedUsage } from './AdvancedUsage.stories'
